+++ src/@ionic-native/plugins/extended-device-information/index.ts
@@ -4,7 +4,11 @@ import { CordovaProperty, IonicNativePlugin, Plugin } from '@ionic-native/core';
 /**
  * @name Extended Device Information
  * @description
- * Retrieves additional device information that is not provided by other plugins
+ * Retrieves additional device information from the Device Hardware
+ *  - memory
+ *  - cpumhz
+ *  - totalstorage
+ *  - freestorage
  *
  * @usage
  * ```typescript
@@ -22,12 +26,12 @@ import { CordovaProperty, IonicNativePlugin, Plugin } from '@ionic-native/core';
   pluginName: 'ExtendedDeviceInformation',
   plugin: 'cordova-plugin-extended-device-information',
   pluginRef: 'extended-device-information',
-  repo: 'https://github.com/danielehrhardt/cordova-plugin-extended-device-information',
+  repo:
+    'https://github.com/danielehrhardt/cordova-plugin-extended-device-information',
   platforms: ['Android']
 })
 @Injectable()
 export class ExtendedDeviceInformation extends IonicNativePlugin {
-
   /**
    * Get the device's memory size
    */
@@ -46,4 +50,9 @@ export class ExtendedDeviceInformation extends IonicNativePlugin {
   @CordovaProperty()
   totalstorage: string;
 
+  /**
+   * Get the total storage
+   */
+  @CordovaProperty
+  freestorage: number;
 }