+++ src/@ionic-native/plugins/network/index.ts
@@ -88,7 +88,8 @@ export class Network extends IonicNativePlugin {
    */
   @Cordova({
     eventObservable: true,
-    event: 'offline'
+    event: 'offline',
+    element: document
   })
   onDisconnect(): Observable<any> {
     return;
@@ -100,7 +101,8 @@ export class Network extends IonicNativePlugin {
    */
   @Cordova({
     eventObservable: true,
-    event: 'online'
+    event: 'online',
+    element: document
   })
   onConnect(): Observable<any> {
     return;