+++ src/@ionic-native/plugins/nfc/index.ts
@@ -341,27 +341,13 @@ export class Ndef extends IonicNativePlugin {
   }
 
   @Cordova({ sync: true })
-  textRecord(
-    text: string,
-    languageCode: string,
-    id: number[] | string
-  ): NdefRecord {
-    return;
-  }
+  textRecord(text: string, languageCode?: string, id?: number[] | string): NdefRecord { return; }
 
   @Cordova({ sync: true })
-  uriRecord(uri: string, id: number[] | string): NdefRecord {
-    return;
-  }
+  uriRecord(uri: string, id?: number[] | string): NdefRecord { return; }
 
   @Cordova({ sync: true })
-  absoluteUriRecord(
-    uri: string,
-    payload: number[] | string,
-    id: number[] | string
-  ): NdefRecord {
-    return;
-  }
+  absoluteUriRecord(uri: string, payload: number[] | string, id?: number[] | string): NdefRecord { return; }
 
   @Cordova({ sync: true })
   mimeMediaRecord(mimeType: string, payload: string): NdefRecord {